Adição de paginação de instrumentos e melhorias no auth

+++ server/src/application/users/use-cases/delete-user.ts
@@ -1,14 +1,21 @@
+import { Role } from '../../../../generated/prisma';
 import { UserEntity } from '../../../domain/entities/user.entity';
 import { UserRepository } from '../../../domain/repositories/user.repository';
+import { UserResponseDto } from '../../../presentation/users/dtos/user-response.dto';
+import { CustomError } from '../../../shared/errors/custom.error';
 
 export interface DeleteUserUseCase {
-  execute(id: string): Promise<UserEntity>;
+  execute(id: string, loggedUser?: UserEntity): Promise<UserResponseDto>;
 }
 
 export class DeleteUser implements DeleteUserUseCase {
   constructor(private readonly repository: UserRepository) {}
 
-  execute(id: string): Promise<UserEntity> {
-    return this.repository.deleteById(id);
+  async execute(id: string, loggedUser?: UserEntity): Promise<UserResponseDto> {
+    if (loggedUser?.role !== Role.ADMIN_ROLE && loggedUser?.id !== id)
+      throw CustomError.forbidden('Ação não permitida');
+
+    const user = await this.repository.deleteById(id);
+    return UserResponseDto.fromEntity(user);
   }
 }
